+++ click_extra/__init__.py
@@ -279,7 +279,7 @@ __all__ = [
 """
 
 
-__version__ = "7.15.1"
+__version__ = "7.15.2.dev0"
 __git_branch__ = ""
 __git_date__ = ""
 __git_long_hash__ = ""